refactor: place pdf testing into their own separate files
