+++ tests/test_recipeorganizer.py
@@ -695,5 +695,116 @@ def test_register_function_introspects_through_a_decorator(input_function):
     assert actual_argnames == expected_argnames
 
 
+def _make_iterpars_tree():
+    """Build a small hierarchy for iterPars tests."""
+    root = RecipeContainer("root")
+    root._containers = {}
+    root._manage(root._containers)
+
+    root_biso = Parameter("Biso", 10)
+    root._add_object(root_biso, root._parameters)
+
+    ni0 = RecipeContainer("Ni0")
+    ni0_biso = Parameter("Biso", 20)
+    ni0_uiso = Parameter("Uiso", 30)
+    ni0._add_object(ni0_biso, ni0._parameters)
+    ni0._add_object(ni0_uiso, ni0._parameters)
+
+    ni1 = RecipeContainer("Ni1")
+    ni1_biso = Parameter("Biso", 40)
+    ni1._add_object(ni1_biso, ni1._parameters)
+
+    o0 = RecipeContainer("O0")
+    o0_biso = Parameter("Biso", 50)
+    o0._add_object(o0_biso, o0._parameters)
+
+    root._add_object(ni0, root._containers)
+    root._add_object(ni1, root._containers)
+    root._add_object(o0, root._containers)
+
+    return {
+        "root": root,
+        "root_biso": root_biso,
+        "ni0": ni0,
+        "ni0_biso": ni0_biso,
+        "ni0_uiso": ni0_uiso,
+        "ni1": ni1,
+        "ni1_biso": ni1_biso,
+        "o0": o0,
+        "o0_biso": o0_biso,
+    }
+
+
+@pytest.mark.parametrize(
+    ("pattern", "kwargs", "expected_values"),
+    [
+        # C1: Match leaf parameter names without fullnames.
+        # Expected: all Biso parameters in the hierarchy are returned.
+        (r"^Biso$", {}, [10, 20, 40, 50]),
+        # C2: Match hierarchical names without fullnames.
+        # Expected: no leaf names match the hierarchical pattern.
+        (r"^Ni\d+\.Biso$", {}, []),
+        # C3: Match hierarchical names with fullnames enabled.
+        # Expected: matching Ni Biso parameters are returned.
+        (r"^Ni\d+\.Biso$", {"fullnames": True}, [20, 40]),
+        # C4: Match one hierarchical Uiso name.
+        # Expected: only Ni0.Uiso is returned.
+        (r"^Ni0\.Uiso$", {"fullnames": True}, [30]),
+        # C5: Match one hierarchical Biso name outside Ni containers.
+        # Expected: only O0.Biso is returned.
+        (r"^O0\.Biso$", {"fullnames": True}, [50]),
+        # C6: Disable recursion while matching child fullnames.
+        # Expected: no child parameters are returned.
+        (r"^Ni\d+\.Biso$", {"fullnames": True, "recurse": False}, []),
+        # C7: Disable recursion while matching root fullname.
+        # Expected: only the root-level Biso parameter is returned.
+        (r"^Biso$", {"fullnames": True, "recurse": False}, [10]),
+    ],
+)
+def test_iterpars_fullname_matching(pattern, kwargs, expected_values):
+    """Verify leaf-name and fullname matching in
+    iterate_over_parameters."""
+    objs = _make_iterpars_tree()
+    root = objs["root"]
+
+    actual_values = [
+        parameter.value
+        for parameter in root.iterate_over_parameters(pattern, **kwargs)
+    ]
+
+    assert actual_values == expected_values
+
+
+@pytest.mark.parametrize(
+    ("pattern", "expected_name"),
+    [
+        # C1: Match Biso relative to the called Ni0 container.
+        # Expected: Ni0.Biso is returned without the Ni0 prefix.
+        (r"^Biso$", ["Biso"]),
+        # C2: Match Uiso relative to the called Ni0 container.
+        # Expected: Ni0.Uiso is returned without the Ni0 prefix.
+        (r"^Uiso$", ["Uiso"]),
+        # C3: Match with the parent container prefix from inside Ni0.
+        # Expected: no parameter is returned because fullnames are relative
+        # to the container on which iterate_over_parameters is called.
+        (r"^Ni0\.Biso$", []),
+    ],
+)
+def test_iterpars_fullnames_are_relative_to_called_container(
+    pattern,
+    expected_name,
+):
+    """Verify fullname matching is relative to the called container."""
+    objs = _make_iterpars_tree()
+    ni0 = objs["ni0"]
+
+    actual_name = [
+        parameter.name
+        for parameter in ni0.iterate_over_parameters(pattern, fullnames=True)
+    ]
+
+    assert actual_name == expected_name
+
+
 if __name__ == "__main__":
     unittest.main()
